chore: organise app.py file, introduced two new files: Nodefinder.py and constants.py
