+++ src/constants.py
@@ -1,0 +1,2 @@
+# Default map centre (web_mercator coordinates)
+DEFAULT_CENTRE = [-356034, 7258806]
